+++ Sources/MapboxCoreNavigation/RouteController.swift
@@ -48,9 +48,9 @@ open class RouteController: NSObject {
     public unowned var dataSource: RouterDataSource
     
     /**
-     Routing source type used to create the route.
+     `NavigationProvider`, used to create route.
      */
-    public var routingSource: NavigationRouter.RouterSource
+    public var routingProvider: NavigationProvider
     
     public var route: Route {
         return routeProgress.route
@@ -186,7 +186,7 @@ open class RouteController: NSObject {
     
     var didFindFasterRoute = false
     
-    var routeTask: NavigationRouter.RoutingRequest?
+    var routeTask: NavigationProviderRequest?
     
     // MARK: Navigating
     
@@ -520,8 +520,12 @@ open class RouteController: NSObject {
     
     // MARK: Handling Lifecycle
     
-    required public init(alongRouteAtIndex routeIndex: Int, in routeResponse: RouteResponse, options: RouteOptions, routingSource: NavigationRouter.RouterSource = .hybrid, dataSource source: RouterDataSource) {
-        self.routingSource = routingSource
+    required public init(alongRouteAtIndex routeIndex: Int,
+                         in routeResponse: RouteResponse,
+                         options: RouteOptions,
+                         routingProvider: NavigationProvider,
+                         dataSource source: RouterDataSource) {
+        self.routingProvider = routingProvider
         self.indexedRouteResponse = .init(routeResponse: routeResponse, routeIndex: routeIndex)
         self.routeProgress = RouteProgress(route: routeResponse.routes![routeIndex], options: options)
         self.dataSource = source
@@ -679,8 +683,7 @@ extension RouteController: Router {
                      routeOptions: RouteOptions?,
                      isProactive: Bool,
                      completion: ((Bool) -> Void)?) {
-        guard let routes = indexedRouteResponse.routeResponse.routes,
-              routes.count > indexedRouteResponse.routeIndex else {
+        guard let route = indexedRouteResponse.selectedRoute else {
             preconditionFailure("`indexedRouteResponse` does not contain route for index `\(indexedRouteResponse.routeIndex)` when updating route.")
         }
         if shouldStartNewBillingSession(for: route, routeOptions: routeOptions) {