+++ Tests/MapboxCoreNavigationTests/RouteControllerTests.swift
@@ -29,7 +29,7 @@ class RouteControllerTests: TestCase {
         let locationManager = ReplayLocationManager(locations: locations)
         replayManager = locationManager
         let equivalentRouteOptions = NavigationRouteOptions(navigationMatchOptions: options)
-        let routeController = RouteController(alongRouteAtIndex: 0, in: routeResponse, options: equivalentRouteOptions, dataSource: self)
+        let routeController = RouteController(alongRouteAtIndex: 0, in: routeResponse, options: equivalentRouteOptions, routingProvider: NavigationRouter(.offline), dataSource: self)
         locationManager.delegate = routeController
         let routerDelegateSpy = RouterDelegateSpy()
         routeController.delegate = routerDelegateSpy
@@ -75,7 +75,7 @@ class RouteControllerTests: TestCase {
         let routeController = RouteController(alongRouteAtIndex: 0,
                                               in: routeResponse,
                                               options: navOptions,
-                                              routingSource: .offline,
+                                              routingProvider: NavigationRouter(.offline),
                                               dataSource: self)
 
         let routerDelegateSpy = RouterDelegateSpy()
@@ -112,26 +112,12 @@ class RouteControllerTests: TestCase {
         }
         
         NavigationRouter.__testRoutesStub = { (options, completionHandler) in
+            DispatchQueue.main.async {
+                completionHandler(Directions.Session(options, .mocked),
+                                  .success(routeResponse))
+            }
             calculateRouteCalled.fulfill()
-            let currentCoordinate = locationManager.location!.coordinate
-            
-            let originWaypoint = Waypoint(coordinate: currentCoordinate)
-            let destinationWaypoint = Waypoint(coordinate: destination)
-            
-            let waypoints = [
-                originWaypoint,
-                destinationWaypoint
-            ]
-            
-            completionHandler(Directions.Session(options, DirectionsCredentials()),
-                              .success(RouteResponse(httpResponse: nil,
-                                                     identifier: nil,
-                                                     routes: [Fixture.route(between: currentCoordinate,
-                                                                            and: destination).route],
-                                                     waypoints: waypoints,
-                                                     options: .route(options),
-                                                     credentials: .mocked)))
-            return 0
+            return nil
         }
 
         let replayFinished = expectation(description: "Replay Finished")