+++ CHANGELOG.md
@@ -59,6 +59,10 @@
 * Renamed the `Locale.usesMetric` property to `Locale.measuresDistancesInMetricUnits`. `Locale.usesMetric` is still available but deprecated. ([#3547](https://github.com/mapbox/mapbox-navigation-ios/pull/3547))
 * Fixed an issue where the user interface did not necessarily display distances in the same units as the route by default. `NavigationRouteOptions` and `NavigationMatchOptions` now set `DirectionsOptions.distanceMeasurementSystem` to a default value matching the `NavigationSettings.distanceUnit` property. ([#3541](https://github.com/mapbox/mapbox-navigation-ios/pull/3541))
 
+### Other changes
+
+* Introduced `NavigationProvider` to parameterize routing fetching and refreshing during active guidance sessions.  `Directions.calculateWithCache(options:completionHandler:)` and `Directions.calculateOffline(options:completionHandler)` functionality is replaced by `NavigationRouter`. It is now recommended to use `NavigationRouter` to request or refresh routes instead of `Directions` object but you may also provide your own `NavigationProvider` implementation to `NavigationService`, `RouteController` or `LegacyRouteController`.  ([#3261](https://github.com/mapbox/mapbox-navigation-ios/pull/3261))
+
 ## v2.0.1
 
 * Added the `Notification.Name.didArriveAtWaypoint` constant for notifications posted when the user arrives at a waypoint. ([#3514](https://github.com/mapbox/mapbox-navigation-ios/pull/3514))